+++ CHANGELOG.md
@@ -6,6 +6,7 @@ HEAD
 
 * Drop support for C++98/C++03. Minimum required is C++11.
 * Remove `ARDUINOJSON_NAMESPACE`; use `ArduinoJson` instead.
+* Make string support generic (issue #1807)
 
 v6.20.1 (2023-02-08)
 -------